+++ src/geo/mercator_coordinate.test.ts
@@ -1,4 +1,4 @@
-import LngLat from './/lng_lat_complete';
+import LngLat from './lng_lat';
 import MercatorCoordinate, {mercatorScale} from './mercator_coordinate';
 
 describe('LngLat', () => {
