+++ src/geo/transform.ts
@@ -1,4 +1,4 @@
-import LngLat from './/lng_lat_complete';
+import LngLat from './lng_lat';
 import LngLatBounds from './lng_lat_bounds';
 import MercatorCoordinate, {mercatorXfromLng, mercatorYfromLat, mercatorZfromAltitude} from './mercator_coordinate';
 import Point from '@mapbox/point-geometry';
