+++ src/source/geojson_source.test.ts
@@ -2,7 +2,7 @@ import Tile from './tile';
 import {OverscaledTileID} from './tile_id';
 import GeoJSONSource, {GeoJSONSourceOptions} from './geojson_source';
 import Transform from '../geo/transform';
-import LngLat from '../geo//lng_lat_complete';
+import LngLat from '../geo/lng_lat';
 import {extend} from '../util/util';
 import Dispatcher from '../util/dispatcher';
 import {RequestManager} from '../util/request_manager';
