+++ src/ui/camera.ts
@@ -1,7 +1,7 @@
 import {bindAll, extend, warnOnce, clamp, wrap, ease as defaultEasing, pick} from '../util/util';
 import {number as interpolate} from '../style-spec/util/interpolate';
 import browser from '../util/browser';
-import LngLat from '../geo//lng_lat_complete';
+import LngLat from '../geo/lng_lat';
 import LngLatBounds from '../geo/lng_lat_bounds';
 import Point from '@mapbox/point-geometry';
 import {Event, Evented} from '../util/evented';
@@ -9,7 +9,7 @@ import {Debug} from '../util/debug';
 import Terrain from '../render/terrain';
 
 import type Transform from '../geo/transform';
-import type {LngLatLike} from '../geo//lng_lat';
+import type {LngLatLike} from '../geo/lng_lat_partial';
 import type {LngLatBoundsLike} from '../geo/lng_lat_bounds';
 import type {TaskID} from '../util/task_queue';
 import type {PaddingOptions} from '../geo/edge_insets';
