+++ src/ui/control/geolocate_control.ts
@@ -1,7 +1,7 @@
 import {Event, Evented} from '../../util/evented';
 import DOM from '../../util/dom';
 import {extend, bindAll, warnOnce} from '../../util/util';
-import LngLat from '../../geo//lng_lat_complete';
+import LngLat from '../../geo/lng_lat';
 import Marker from '../marker';
 
 import type Map from '../map';
