+++ src/ui/handler/scroll_zoom.ts
@@ -3,7 +3,7 @@ import DOM from '../../util/dom';
 import {ease as _ease, bindAll, bezier} from '../../util/util';
 import browser from '../../util/browser';
 import {number as interpolate} from '../../style-spec/util/interpolate';
-import LngLat from '../../geo//lng_lat_complete';
+import LngLat from '../../geo/lng_lat';
 
 import type Map from '../map';
 import type HandlerManager from '../handler_manager';
