+++ src/ui/marker.test.ts
@@ -1,7 +1,7 @@
 import {createMap as globalCreateMap, beforeMapTest} from '../util/test/util';
 import Marker from './marker';
 import Popup from './popup';
-import LngLat from '../geo//lng_lat_complete';
+import LngLat from '../geo/lng_lat';
 import Point from '@mapbox/point-geometry';
 import simulate from '../../test/unit/lib/simulate_interaction';
 import type Terrain from '../render/terrain';
