+++ src/ui/marker.ts
@@ -1,5 +1,5 @@
 import DOM from '../util/dom';
-import LngLat from '../geo//lng_lat_complete';
+import LngLat from '../geo/lng_lat';
 import Point from '@mapbox/point-geometry';
 import smartWrap from '../util/smart_wrap';
 import {bindAll, extend} from '../util/util';
@@ -8,7 +8,7 @@ import type {PositionAnchor} from './anchor';
 import {Event, Evented} from '../util/evented';
 import type Map from './map';
 import Popup, {Offset} from './popup';
-import type {LngLatLike} from '../geo//lng_lat';
+import type {LngLatLike} from '../geo/lng_lat_partial';
 import type {MapMouseEvent, MapTouchEvent} from './events';
 import type {PointLike} from './camera';
