+++ src/ui/popup.test.ts
@@ -1,6 +1,6 @@
 import {createMap as globalCreateMap, beforeMapTest} from '../util/test/util';
 import Popup, {Offset} from './popup';
-import LngLat from '../geo//lng_lat_complete';
+import LngLat from '../geo/lng_lat';
 import Point from '@mapbox/point-geometry';
 import simulate from '../../test/unit/lib/simulate_interaction';
 import {PositionAnchor} from './anchor';
