+++ src/ui/popup.ts
@@ -2,14 +2,14 @@ import {extend, bindAll} from '../util/util';
 import {Event, Evented} from '../util/evented';
 import {MapMouseEvent} from '../ui/events';
 import DOM from '../util/dom';
-import LngLat from '../geo//lng_lat_complete';
+import LngLat from '../geo/lng_lat';
 import Point from '@mapbox/point-geometry';
 import smartWrap from '../util/smart_wrap';
 import {anchorTranslate, applyAnchorClass} from './anchor';
 
 import type {PositionAnchor} from './anchor';
 import type Map from './map';
-import type {LngLatLike} from '../geo//lng_lat';
+import type {LngLatLike} from '../geo/lng_lat_partial';
 import type {PointLike} from './camera';
 
 const defaultOptions = {
